Make panning and zooming simpler and cross browser compatible.
By using a delta calculation the panning becomes much easier to handle
and understand. See `addPanning`.

Zooming is now done via a more robust value that is supplied via
`jointjs`. This new value just has the direction and we now multiply
that with a hardcoded factor. The result is robust zooming on firefox
and chrome.

Both these implementation now no longer use deprecated or experimental
features.

-- closes #65