Try not be smart about types.
The commit ce6fc74 introduced logic that would try to construct
a new generic type from all previously connected ports.

For doing so it would not use the actual type but try to reconstruct it
from the current port e.g. `this`. See the method `getConnectedType`.

This method however is not able to do that correctly and would often fail
thus resulting in the error `types not unifiable`.

The questions I was asking myself:
- why would we want to try to reconstruct a type and unify it?
- what was the original intention behind the implementation? - sadly that
  is not clear from the code, the commit or the comments :-/
- wouldn't it be easier to just leave everything as it is and not
  surprise the user?
- how can we avoid throwing the error and not freezing the editor?

In order to lower the surprise factor I chose to remove most of the
original logic and simply disconnect two ports if their types are not
compatible.

The optimal solution in my opinion would be to check if the type change
can be propagated through all connected ports - both ways - and only
than apply the change w/o disconnecting anything. If at any point along
the connections/edges we encounter an incompatible type, cut the connection
between the root and the rest of the nodes.

But you know what - that is so far out of scope right now!