This is the cleanup from the panning and zoom refactoring (#113)
* Make panning and zooming simpler and cross browser compatible.

By using a delta calculation the panning becomes much easier to handle
and understand. See `addPanning`.

Zooming is now done via a more robust value that is supplied via
`jointjs`. This new value just has the direction and we now multiply
that with a hardcoded factor. The result is robust zooming on firefox
and chrome.

Both these implementation now no longer use deprecated or experimental
features.

-- closes #65

* Remove unused variables

* use better unused var names

* Use smooth step function for zooming

* Use normal linear scrolling just a bit slower

* Remove a lot of unneeded cruft and move getters to the top

* Remove debugger statment

* Remove unused variable

* It is actually the other way around