Fix an bug where the viewFrame size would always be 600px.
That is because the defaul PaperView size is 600px.
When the viewFrame gets created it did not have a height, which led
to the viewFrame always being the size of the default PaperView.

Now the viewFrame always uses all it's available height, which is
the height the parent.

This enables containing the editor correctly inside an element.

//cc @td5r